zabbix-genai-insights/requirements.txt
google-generativeai
openai
requests
python-dotenv
